// ClientApp/generated/UserEditModel.ts
/*  GENERTAED automatically using Models.tst. DO NOT CHANGE HERE */
import { WorkspaceUserModel } from './WorkspaceUserModel';
//Generated from class components.permissions.UserEditModel
export interface UserEditModel {
  workspaceUser: WorkspaceUserModel;
  email: string;
  prefferedName?: string;
  isAdmin?: boolean;
  inviteToJoin: boolean;
}